requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pywebpush==1.14.1
cryptography>=42.0.0
python-telegram-bot==21.5
itsdangerous==2.2.0
psycopg2-binary==2.9.9

